Ajout de tests sur le parseur.
Parseur semble fini, créer des cas de test supplémentaire pour s'en
assurer.

+++ src/main/java/parseurcroustillant/view/ExclusiveChoiceQuestionPanel.java
@@ -30,7 +30,7 @@ public class ExclusiveChoiceQuestionPanel extends QuestionPanel {
 	}
 
 	@Override
-	public int processSubmit() {
+	public float processSubmit() {
 		for(ButtonGroup g : mButtonGroups) {
 			//g.getElements()
 		}
