+++ src/main/java/parseurcroustillant/view/FillInTheBlankQuestionPanel.java
@@ -17,7 +17,7 @@ public class FillInTheBlankQuestionPanel extends QuestionPanel {
 	}
 
 	@Override
-	public int processSubmit() {
+	public float processSubmit() {
 		//TODO
 		return 0;
 	}
